back/requirements.txt
asgiref==3.8.1
axios==0.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.16
django-allauth==0.59.0
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
idna==3.10
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
